ProductServiceImplementation

+++ src/main/java/entity/Cloth.java
@@ -6,7 +6,7 @@ public class Cloth extends Product {
     String size;
     String material;
 
-    public Cloth(int id, String productName, Float price, Float weight, String color, Integer productCount, String size, String material) {
+    public Cloth(long id, String productName, Float price, Float weight, String color, Integer productCount, String size, String material) {
         super(id, productName, price, weight, color, productCount);
         this.size = size;
         this.material = material;
